Move `globset` forked code to a new file, improve the license text
fix how directory match regexes are built to exclude suffixes.
